argh

+++ src/board.h
@@ -80,10 +80,10 @@ typedef enum
 //	  GPS_MTK19 = 3,
 //} GPSHardware;
 
-typedef void     (* sensorInitFuncPtr)(void);               // sensor init prototype
-typedef void     (* sensorReadFuncPtr)(int16_t *data);      // sensor read and align prototype
-typedef float    (* baroCalculateFuncPtr)(void);            // baro calculation (returns altitude in cm based on static data collected)
-typedef void     (* uartReceiveCallbackPtr)(uint16_t data); // used by uart2 driver to return frames to app
+typedef void (* sensorInitFuncPtr)(void);                   // sensor init prototype
+typedef void (* sensorReadFuncPtr)(int16_t *data);          // sensor read and align prototype
+typedef int32_t (* baroCalculateFuncPtr)(void);             // baro calculation (returns altitude in cm based on static data collected)
+typedef void (* uartReceiveCallbackPtr)(uint16_t data);     // used by uart2 driver to return frames to app
 typedef uint16_t (* rcReadRawDataPtr)(uint8_t chan);        // used by receiver driver to return channel data
 
 typedef struct sensor_t
